feat:updata app

+++ service/api/app/internal/repo/mempool/api.go
@@ -17,7 +17,7 @@ import (
 
 type Mempool interface {
 	GetMempoolTxs(offset int64, limit int64) (mempoolTx []*mempoolModel.MempoolTx, err error)
-	GetMempoolTxsTotalCount() (count int64, err error)
+	GetMempoolTxsTotalCount() (count uint32, err error)
 	GetMempoolTxsTotalCountByAccountIndex(accountIndex int64) (count int64, err error)
 	//GetMempoolTxsListByAccountIndex(accountIndex int64, limit int64, offset int64) (mempoolTx []*types.Tx, err error)
 	//GetMempoolTxsTotalCountByPublicKey(pk string) (mempoolTx []*types.Tx, err error)
